fix: viirs path changes
As of around 2024.06.15 the VNP10A1F.001 product does not seem to have any granules.  It may go back further in date.  Looks like there is a new version VNP10A1F.002 that has a bunch of granules, so have modified the code to switch to that product if the date is after June 15, 2024.