fix: archive days change (#107)
* fix: change the number of days to look back when archiving data

* fix: don't need this file

* fix: update the lock file

* fix: add debug config for the viirs get dates

* docs: update documentation

* fix: tweak the way the script works so that the cli options are properly configured

* feat: starting to work on a script to fill in missing dates

* fix: add tmp mamba env to ignore list

* fix: adjust the call to the get_available_data.py script

* fix: viirs path changes

As of around 2024.06.15 the VNP10A1F.001 product does not seem to have any granules.  It may go back further in date.  Looks like there is a new version VNP10A1F.002 that has a bunch of granules, so have modified the code to switch to that product if the date is after June 15, 2024.

* feat: complete the batch run script

* fix:  separate the click methods from the methods that do the work

This allows the batchrun script to call the methods that are glued to the @click cli

* fix:  commenting out duplications

* fix: minor updates to docs

* fix: commenting out temp code used for debugging data holes on earthdata source

* docs: update the docs